add checkEmailExists middleware
